+++ CHANGELOG.md
@@ -3,6 +3,22 @@ CHANGELOG
 
 ## HEAD (Unreleased)
 * Add support for `arn:aws:iam::aws:policy/EC2InstanceProfileForImageBuilder` in `aws.iam.ManagedPolicy`
+* `aws.iam.CallbackFunction` will now default to the following policies if no policies are specified by the user: 
+  * LambdaFullAccess
+  * CloudWatchFullAccess
+  * CloudWatchEventsFullAccess
+  * AmazonS3FullAccess
+  * AmazonDynamoDBFullAccess
+  * AmazonSQSFullAccess
+  * AmazonKinesisFullAccess
+  * AmazonCognitoPowerUser
+  * AWSXrayWriteOnlyAccess
+* `aws.kinesis.Stream.onEvent` will now default to the following policies:
+  * AWSLambdaKinesisExecutionRole
+  * AmazonKinesisFullAccess
+  * CloudWatchFullAccess
+  * CloudWatchEventsFullAccess
+  * LambdaFullAccess
 
 ---